+++ test/plottingTest.jl
@@ -6,7 +6,7 @@ using Manifolds
 
 ##
 
-# struct TestInferenceVariable1 <: InferenceVariable end
+# struct TestInferenceVariable1 <: VariableStateType end
 @defVariable TestInferenceVariable1 Euclidean(1) [0.0;]
 
 # Now make a complex graph for connectivity tests